Added options to download datasets and save / load the trained model.
Added functions that help download and unzip zip file.
This is in order to help with downloading datasets and then unzip them, so that
the model knows exactly where to load the train and test data from.

Added functions that help load image data to train the model on.
To train on images the model now can load image data in form of a NxM matrix.
The functions flattens the data and then normalizes it. Lastly it shuffels the
training data to help with generalization.

Added functions that helps to get / set the parameters of the model to a new model.
These help with saving the parameters of a currently training / trained model then
load it into a different model. These are also used when saving / loading the model
to a file.

Added functions to help to save and load the entire model object.
This required to add additional functionalities to the already
existing classes, but it helps to make quick-saves of models
that are training, and allows the model to be loaded from a
file to be used for evaluations and predictions.

+++ include/Loss/LossBinaryCrossEntropy.h
@@ -1,7 +1,7 @@
 #ifndef __LOSS_BINARY_CROSS_ENTROPY_H__
 #define __LOSS_BINARY_CROSS_ENTROPY_H__
 
-#include "Loss/Loss.h"
+#include "Loss.h"
 
 namespace NEURAL_NETWORK
 {
@@ -12,12 +12,11 @@ namespace NEURAL_NETWORK
 		~LossBinaryCrossEntropy() = default;
 
 		void forward(const Eigen::MatrixXd& predictions,
-					 const Eigen::MatrixXi& targets) override;
+					 const Eigen::MatrixXd& targets) override;
 
 		void backward(const Eigen::MatrixXd& d_values,
-					  const Eigen::MatrixXi& targets) override;
+					  const Eigen::MatrixXd& targets) override;
 	};
-
 } // namespace NEURAL_NETWORK
 
 #endif // __LOSS_BINARY_CROSS_ENTROPY_H__
